moved c++ files to cpp; added cmake build
