+++ cpp/cmake/FindPackages.cmake
@@ -1,0 +1,5 @@
+# In this CMake file, we will find all required packages
+
+find_package(Libint2 2.0.3 REQUIRED)
+
+find_package(MKL REQUIRED)
