+++ cpp/src/calc.cpp
