+++ cpp/src/calc.h
