+++ cpp/src/cartpure.cpp
