+++ cpp/src/cartpure.h
