+++ cpp/src/common.cpp
