+++ cpp/src/common.h
