+++ cpp/src/grids/fns.cpp
@@ -28,8 +28,8 @@
 #include <cstdlib>
 #include <cstring>
 #include <stdexcept>
-#include "boys.h"
-#include "cartpure.h"
+#include "../twos/boys.h"
+#include "../cartpure.h"
 #include "fns.h"
 
 /*
