+++ cpp/src/grids/iter_pow1.cpp
@@ -19,7 +19,7 @@
 //--
 
 
-#include "iter_pow.h"
+#include "iter_pow1.h"
 
 int iter_pow1_inc(long *n) {
   // Modify the indexes in place as to move to the next combination of powers
@@ -68,45 +68,3 @@ int IterPow1::inc() {
   return result;
 }
 
-/*
-
-  IterPow2
-
-*/
-
-
-void IterPow2::reset(long shell_type0, long shell_type1) {
-  this->shell_type0 = shell_type0;
-  this->shell_type1 = shell_type1;
-  n0[0] = shell_type0;
-  n0[1] = 0;
-  n0[2] = 0;
-  n1[0] = shell_type1;
-  n1[1] = 0;
-  n1[2] = 0;
-  ibasis0 = 0;
-  ibasis1 = 0;
-  offset = 0;
-}
-
-int IterPow2::inc() {
-  // Increment indexes of shell 1
-  int result;
-  result = iter_pow1_inc(n1);
-  if (result) {
-    offset++;
-    ibasis1++;
-  } else {
-    // Increment indexes of shell 0
-    result = iter_pow1_inc(n0);
-    ibasis1 = 0;
-    if (result) {
-      offset++;
-      ibasis0++;
-    } else {
-      offset = 0;
-      ibasis0 = 0;
-    }
-  }
-  return result;
-}
