+++ cpp/src/ones/nucpot.cpp
