+++ cpp/src/ones/nucpot.h
