+++ cpp/src/sparse/CMakeLists.txt
@@ -1,0 +1,7 @@
+target_sources(${LIBRARY_NAME}
+        PUBLIC
+            "${CMAKE_CURRENT_LIST_DIR}/cholesky.cpp"
+            "${CMAKE_CURRENT_LIST_DIR}/cholesky.h"
+        )
+
+target_include_directories(${LIBRARY_NAME} PUBLIC "${CMAKE_CURRENT_LIST_DIR}")
