+++ cpp/src/twos/boys.h
@@ -26,7 +26,7 @@
 #ifndef GBASIS_BOYS_H_
 #define GBASIS_BOYS_H_
 
-#include "common.h"
+#include "../common.h"
 
 #define BOYS_MAX_M 4*MAX_SHELL_TYPE
