+++ cpp/src/twos/gbw.cpp
