+++ cpp/src/twos/gbw.h
@@ -26,8 +26,8 @@
 #ifndef GBASIS_GBW_H_
 #define GBASIS_GBW_H_
 
-#include "gbasis.h"
-#include "ints.h"
+#include "../gbasis.h"
+#include "ints4.h"
 
 /**
     @brief
